+++ src/full/Agda/Syntax/Internal.hs
@@ -56,6 +56,19 @@ import Agda.Utils.Impossible
 -- * Function type domain
 ---------------------------------------------------------------------------
 
+data RewDom' t = RewDom
+  { rewDomEq  :: LocalEquation' t
+    -- ^ Elaborated "@rewrite" equation
+  , rewDomRew :: Maybe RewriteRule
+    -- ^ "@rewrite" equation transformed into a directed rewrite rule.
+    --
+    -- @Nothing@ iff invalidated by a substitution. If we are checking
+    -- against an "@rewrite" domain, this is fine, but if we are inside an "@rewrite"
+    -- context, this is probably an internal error.
+  } deriving (Show, Generic)
+
+type RewDom = RewDom' Term
+
 -- | Similar to 'Arg', but we need to distinguish
 --   an irrelevance annotation in a function domain
 --   (the domain itself is not irrelevant!)
@@ -76,17 +89,33 @@ data Dom' t e = Dom
   , domIsFinite :: Bool
     -- ^ Is this a Π-type (False), or a partial type (True)?
   , domTactic :: Maybe t        -- ^ "@tactic e".
-  , domRewrite :: Maybe RewriteRule
+  , rewDom    :: Maybe (RewDom' t)
+    -- ^ Elaborated "@rewrite" equation
+    --
+    -- Will only be present if domain annotated with "@rewrite" (@annRewrite@
+    -- is @IsRewrite@) AND the type successfully elaborated into a rewrite rule.
   , unDom     :: e
   } deriving (Show, Functor, Foldable, Traversable)
 
 type Dom = Dom' Term
+
+domEq :: Dom' t e -> Maybe (LocalEquation' t)
+domEq = fmap rewDomEq . rewDom
+
+-- | Is this Dom annotated as a local rewrite rule and if so, has the rewrite
+--   been invalidated due to a substitution?
+invalidRew :: Dom' t e -> Bool
+invalidRew Dom { rewDom = Just (RewDom { rewDomRew = Nothing }) } = True
+invalidRew _                                                      = False
 
 instance Decoration (Dom' t) where
   traverseF f (Dom ai x t b r a) = Dom ai x t b r <$> f a
 
 instance HasRange a => HasRange (Dom' t a) where
   getRange = getRange . unDom
+
+instance KillRange t => KillRange (RewDom' t) where
+  killRange (RewDom eq rew) = killRangeN RewDom eq rew
 
 instance (KillRange t, KillRange a) => KillRange (Dom' t a) where
   killRange (Dom info x t b r a) = killRangeN Dom info x t b r a
@@ -112,6 +141,9 @@ instance LensLock (Dom' t e) where
 instance LensRewriteAnn (Dom' t e) where
   getRewriteAnn = getRewriteAnn . getArgInfo
   setRewriteAnn = mapRewriteAnn . setRewriteAnn
+
+instance LensLocalEquation (Dom e) where
+  getLocalEq = domEq
 
 -- The other lenses are defined through LensArgInfo
 
@@ -139,8 +171,11 @@ namedArgFromDom Dom{domInfo = i, domName = s, unDom = a} = Arg i $ Named s a
 -- However, this causes problems with instance resolution in several places.
 -- often for class AddContext.
 
+domFromArgRew :: Maybe RewDom -> Arg a -> Dom a
+domFromArgRew rew (Arg i a) = Dom i Nothing False Nothing rew a
+
 domFromArg :: Arg a -> Dom a
-domFromArg (Arg i a) = Dom i Nothing False Nothing Nothing a
+domFromArg = domFromArgRew Nothing
 
 domFromNamedArg :: NamedArg a -> Dom a
 domFromNamedArg (Arg i a) = Dom i (nameOf a) False Nothing Nothing (namedThing a)
@@ -148,8 +183,11 @@ domFromNamedArg (Arg i a) = Dom i (nameOf a) False Nothing Nothing (namedThing a
 defaultDom :: a -> Dom a
 defaultDom = defaultArgDom defaultArgInfo
 
+defaultArgDomRew :: ArgInfo -> Maybe RewDom -> a -> Dom a
+defaultArgDomRew info rew x = domFromArgRew rew (Arg info x)
+
 defaultArgDom :: ArgInfo -> a -> Dom a
-defaultArgDom info x = domFromArg (Arg info x)
+defaultArgDom info = defaultArgDomRew info Nothing
 
 defaultNamedArgDom :: ArgInfo -> String -> a -> Dom a
 defaultNamedArgDom info s x = (defaultArgDom info x) { domName = Just $ WithOrigin Inserted $ unranged s }
@@ -758,6 +796,7 @@ data Substitution' a
 
 type Substitution = Substitution' Term
 type PatternSubstitution = Substitution' DeBruijnPattern
+type Renaming = Substitution' Nat
 
 infixr 4 :#
 
@@ -1279,15 +1318,49 @@ pattern PSSet p = PUniv USSet p
   PType, PSSet, PProp, PInf,
   PSizeUniv, PLockUniv, PLevelUniv, PIntervalUniv #-}
 
+data RewriteHead
+  = RewDefHead QName
+  | RewVarHead Nat
+    -- ^ de Bruijn index of head symbol (excluding rewContext variables)
+  deriving (Show, Generic, Eq)
+
+headToPat :: Nat -> RewriteHead -> PElims -> NLPat
+headToPat _        (RewDefHead f) = PDef f
+headToPat telStart (RewVarHead x) = PBoundVar (x + telStart)
+
+headToTerm :: Nat -> RewriteHead -> Elims -> Term
+headToTerm _        (RewDefHead f) = Def f
+headToTerm telStart (RewVarHead x) = Var (x + telStart)
+
+-- | Undirected equational constraint ("the LHS and RHS
+--   must be convertible in the calling context").
+--   Admits arbitrary substitution.
+data LocalEquation' t = LocalEquation
+  { lEqContext :: Tele (Dom' t (Type'' t t))
+  , lEqLHS     :: t
+  , lEqRHS     :: t
+  , lEqType    :: Type'' t t
+  }
+  deriving (Show, Generic)
+
+type LocalEquation = LocalEquation' Term
+
+-- | Directed rewrite rules generic over the type of head symbol.
+--   Generally unstable under substitution.
 data RewriteRule = RewriteRule
   { rewContext :: Telescope
-  , rewHead    :: Int        -- de Bruijn index of head symbol (excluding rewContext variables)
+  , rewHead    :: RewriteHead
   , rewPats    :: PElims     -- patterns (including rewContext variables)
   , rewRHS     :: Term
   , rewType    :: Type
   }
   deriving (Show, Generic)
 
+lrewHasProjectionPattern :: RewriteRule -> Bool
+lrewHasProjectionPattern rew = any (isJust . isProjElim) $ rewPats rew
+
+class LensLocalEquation a where
+  getLocalEq :: a -> Maybe LocalEquation
 
 ---------------------------------------------------------------------------
 -- * Null instances.
@@ -1429,16 +1502,16 @@ instance KillRange Term where
     DontCare mv -> killRangeN DontCare mv
     v@Dummy{}   -> v
 
-instance KillRange Level where
+instance KillRange a => KillRange (Level' a) where
   killRange (Max n as) = killRangeN (Max n) as
 
-instance KillRange PlusLevel where
+instance KillRange a => KillRange (PlusLevel' a) where
   killRange (Plus n l) = killRangeN (Plus n) l
 
-instance (KillRange a) => KillRange (Type' a) where
+instance (KillRange a, KillRange b) => KillRange (Type'' a b) where
   killRange (El s v) = killRangeN El s v
 
-instance KillRange Sort where
+instance KillRange a => KillRange (Sort' a) where
   killRange = \case
     Inf u n    -> Inf u n
     SizeUniv   -> SizeUniv
@@ -1521,6 +1594,16 @@ instance KillRange NLPSort where
   killRange PLockUniv = PLockUniv
   killRange PLevelUniv = PLevelUniv
   killRange PIntervalUniv = PIntervalUniv
+
+instance KillRange RewriteHead where
+  killRange (RewVarHead a) =
+    killRangeN RewVarHead a
+  killRange (RewDefHead a) =
+    killRangeN RewDefHead a
+
+instance KillRange t => KillRange (LocalEquation' t) where
+  killRange (LocalEquation a b c d) =
+    killRangeN LocalEquation a b c d
 
 instance KillRange RewriteRule where
   killRange (RewriteRule a b c d e) =
@@ -1764,4 +1847,7 @@ instance NFData DefSing
 instance NFData NLPat
 instance NFData NLPType
 instance NFData NLPSort
+instance NFData RewriteHead
+instance NFData LocalEquation
 instance NFData RewriteRule
+instance NFData RewDom