+++ src/full/Agda/TypeChecking/Monad/Constraints.hs
@@ -222,6 +222,7 @@ isBlockingConstraint = \case
   CheckType{}           -> True
   CheckLockedVars{}     -> True
   UsableAtModality{}    -> True
+  RewConstraint{}       -> True
 
 -- | Start solving constraints
 nowSolvingConstraints :: MonadTCEnv m => m a -> m a