+++ src/full/Agda/TypeChecking/Pretty/Call.hs
@@ -227,6 +227,12 @@ instance PrettyTCM Call where
         , fsep (pwords "must be equal, since" ++ [prettyTCM fn] ++ pwords "could reduce to either.")
         ]
 
+    CheckLocalRewriteConstraint e c -> vcat
+      [ fsep $ pwords "when checking the following local rewrite rule constraint:"
+      , nest 2 $ prettyTCM e
+      , prettyTCM c
+      ]
+
     where
     hPretty :: MonadPretty m => Arg (Named_ Expr) -> m Doc
     hPretty a = do