+++ src/full/Agda/TypeChecking/Reduce.hs
@@ -281,7 +281,7 @@ instance Instantiate Sort where
       _            -> __IMPOSSIBLE__
     s -> return s
 
-instance (Instantiate t, Instantiate e) => Instantiate (Dom' t e) where
+instance Instantiate e => Instantiate (Dom e) where
     instantiate' (Dom i n b tac rew x) = Dom i n b <$> instantiate' tac <*> instantiate' rew <*> instantiate' x
 
 instance Instantiate a => Instantiate (Closure a) where
@@ -318,6 +318,7 @@ instance Instantiate Constraint where
   instantiate' c@CheckMetaInst{}    = return c
   instantiate' (CheckType t)        = CheckType <$> instantiate' t
   instantiate' (UsableAtModality cc ms mod t) = flip (UsableAtModality cc) mod <$> instantiate' ms <*> instantiate' t
+  instantiate' (RewConstraint e)    = RewConstraint <$> instantiate' e
 
 instance Instantiate CompareAs where
   instantiate' (AsTermsOf a) = AsTermsOf <$> instantiate' a
@@ -340,6 +341,14 @@ instance Instantiate EqualityView where
     <*> instantiate' a
     <*> instantiate' b
 
+instance Instantiate LocalEquation where
+  instantiate' (LocalEquation a b c d) =
+    LocalEquation
+      <$> instantiate' a
+      <*> instantiate' b
+      <*> instantiate' c
+      <*> instantiate' d
+
 instance Instantiate RewriteRule where
   instantiate' (RewriteRule a b c d e) =
     RewriteRule
@@ -348,6 +357,12 @@ instance Instantiate RewriteRule where
       <*> pure c
       <*> instantiate' d
       <*> instantiate' e
+
+instance Instantiate RewDom where
+  instantiate' (RewDom a b) =
+    RewDom
+      <$> instantiate' a
+      <*> instantiate' b
 
 ---------------------------------------------------------------------------
 -- * Reduction to weak head normal form.
@@ -605,10 +620,10 @@ slowReduceTerm v = do
       Level l  -> ifM (SmallSet.member LevelReductions <$> asksTC envAllowedReductions)
                     {- then -} (fmap levelTm <$> reduceB' l)
                     {- else -} done
-      Pi _ _   -> done
-      Lit _    -> done
-      Var _ es  -> iapp es
-      Lam _ _  -> done
+      Pi _ _     -> done
+      Lit _      -> done
+      Var x es   -> reduceIApply (rewriteVarApp x es) es
+      Lam _ _    -> done
       DontCare _ -> done
       Dummy{}    -> done
     where
@@ -629,7 +644,23 @@ slowReduceTerm v = do
               w              -> Con c ci [Apply $ defaultArg w]
       reduceNat v = return v
 
--- Andreas, 2013-03-20 recursive invokations of unfoldCorecursion
+rewriteVarApp :: Nat -> Elims -> ReduceM (Blocked Term)
+rewriteVarApp x es = do
+  r <- rewriteVarAppStep x es
+  case r of
+    NoReduction v    -> return v
+    YesReduction _ v -> reduceB' v
+
+rewriteVarAppStep :: Nat -> Elims -> ReduceM (Reduced (Blocked Term) Term)
+rewriteVarAppStep x es = do
+  rewr <- getAllRewriteRulesForVarHead x
+  when (not $ null rewr) $
+    reportSDoc "rewriting" 30 $
+      "Trying to rewrite variable application" <+> prettyTCM (Var x es)
+  rewrite (NotBlocked ReallyNotBlocked ())
+          (Var x) rewr es
+
+-- Andreas, 2013-03-20 recursive invocations of unfoldCorecursion
 -- need also to instantiate metas, see Issue 826.
 unfoldCorecursionE :: Elim -> ReduceM (Blocked Elim)
 unfoldCorecursionE (Proj o p)           = notBlocked . Proj o <$> getOriginalProjection p
@@ -677,7 +708,7 @@ unfoldDefinitionStep v0 f es =
   {-# SCC "reduceDef" #-} do
   traceSDoc "tc.reduce" 90 ("unfoldDefinitionStep v0" <+> pretty v0) $ do
   info <- getConstInfo f
-  rewr <- instantiateRewriteRules =<< getGlobalRewriteRulesFor f
+  rewr <- getAllRewriteRulesForDefHead f
   allowed <- asksTC envAllowedReductions
   prp <- runBlocked $ isPropM $ defType info
   defOk <- shouldReduceDef f
@@ -694,7 +725,7 @@ unfoldDefinitionStep v0 f es =
         , isIrrelevant info
         , not defOk
         ]
-      copatterns = defCopatternLHS info
+  copatterns <- defCopatternLHS f info
   case def of
     Constructor{conSrcCon = c} -> do
       let hd = Con (c `withRangeOf` f) ConOSystem
@@ -753,7 +784,7 @@ unfoldDefinitionStep v0 f es =
 
     reduceNormalE ::
          Term -> QName -> [MaybeReduced Elim] -> Bool -> [Clause]
-      -> Maybe CompiledClauses -> GlobalRewriteRules
+      -> Maybe CompiledClauses -> RewriteRules
       -> ReduceM (Reduced (Blocked Term) Term)
     reduceNormalE v0 f es dontUnfold def mcc rewr = {-# SCC "reduceNormal" #-} do
       traceSDoc "tc.reduce" 90 ("reduceNormalE v0 =" <+> pretty v0) $ do
@@ -902,12 +933,12 @@ unfoldInlined v = do
 -- | Apply a definition using the compiled clauses, or fall back to
 --   ordinary clauses if no compiled clauses exist.
 appDef_ ::
-     QName -> Term -> [Clause] -> Maybe CompiledClauses -> GlobalRewriteRules
+     QName -> Term -> [Clause] -> Maybe CompiledClauses -> RewriteRules
   -> MaybeReducedArgs -> ReduceM (Reduced (Blocked Term) Term)
 appDef_ f v0 cls mcc rewr args = appDefE_ f v0 cls mcc rewr $ map (fmap Apply) args
 
 appDefE_ ::
-     QName -> Term -> [Clause] -> Maybe CompiledClauses -> GlobalRewriteRules
+     QName -> Term -> [Clause] -> Maybe CompiledClauses -> RewriteRules
   -> MaybeReducedElims -> ReduceM (Reduced (Blocked Term) Term)
 appDefE_ f v0 cls mcc rewr args =
   localTC (\ e -> e { envAppDef = Just f }) $
@@ -918,12 +949,12 @@ appDefE_ f v0 cls mcc rewr args =
 -- | Apply a defined function to it's arguments, using the compiled clauses.
 --   The original term is the first argument applied to the third.
 appDef ::
-     Term -> CompiledClauses -> GlobalRewriteRules -> MaybeReducedArgs
+     Term -> CompiledClauses -> RewriteRules -> MaybeReducedArgs
   -> ReduceM (Reduced (Blocked Term) Term)
 appDef v cc rewr args = appDefE v cc rewr $ map (fmap Apply) args
 
 appDefE ::
-     Term -> CompiledClauses -> GlobalRewriteRules -> MaybeReducedElims
+     Term -> CompiledClauses -> RewriteRules -> MaybeReducedElims
   -> ReduceM (Reduced (Blocked Term) Term)
 appDefE v cc rewr es = do
   traceSDoc "tc.reduce" 90 ("appDefE v = " <+> pretty v) $ do
@@ -934,18 +965,20 @@ appDefE v cc rewr es = do
 
 -- | Apply a defined function to it's arguments, using the original clauses.
 appDef' ::
-     QName -> Term -> [Clause] -> GlobalRewriteRules -> MaybeReducedArgs
+     QName -> Term -> [Clause] -> RewriteRules -> MaybeReducedArgs
   -> ReduceM (Reduced (Blocked Term) Term)
 appDef' f v cls rewr args = appDefE' f v cls rewr $ map (fmap Apply) args
 
-appDefE' :: QName -> Term -> [Clause] -> GlobalRewriteRules -> MaybeReducedElims -> ReduceM (Reduced (Blocked Term) Term)
+appDefE' ::
+     QName -> Term -> [Clause] -> RewriteRules -> MaybeReducedElims
+  -> ReduceM (Reduced (Blocked Term) Term)
 appDefE' f v cls rewr es =
   localTC (\ e -> e { envAppDef = Just f }) $
   appDefE'' v cls rewr es
 
 -- | Expects @'envAppDef' = Just f@ in 'TCEnv' to be able to report @'MissingClauses' f@.
 appDefE'' ::
-     Term -> [Clause] -> GlobalRewriteRules -> MaybeReducedElims
+     Term -> [Clause] -> RewriteRules -> MaybeReducedElims
   -> ReduceM (Reduced (Blocked Term) Term)
 appDefE'' v cls rewr es = traceSDoc "tc.reduce" 90 ("appDefE' v = " <+> pretty v) $ do
   goCls cls $ map ignoreReduced es
@@ -1008,6 +1041,12 @@ instance Reduce Telescope where
 instance Reduce ProblemConstraint where
   reduce' (PConstr p u c) = PConstr p u <$> reduce' c
 
+instance Reduce LocalEquation where
+  reduce' (LocalEquation g t u a) = do
+    g' <- reduce' g
+    (t', u', a') <- addContext g' $ reduce' (t, u, a)
+    return $ LocalEquation g' t' u' a'
+
 instance Reduce Constraint where
   reduce' (ValueCmp cmp t u v) = do
     (t,u,v) <- reduce' (t,u,v)
@@ -1034,6 +1073,7 @@ instance Reduce Constraint where
   reduce' c@CheckMetaInst{}     = return c
   reduce' (CheckType t)         = CheckType <$> reduce' t
   reduce' (UsableAtModality cc ms mod t) = flip (UsableAtModality cc) mod <$> reduce' ms <*> reduce' t
+  reduce' (RewConstraint e)     = RewConstraint <$> reduce' e
 
 instance Reduce CompareAs where
   reduce' (AsTermsOf a) = AsTermsOf <$> reduce' a
@@ -1175,6 +1215,12 @@ instance (Subst a, Simplify a) => Simplify (Tele a) where
 instance Simplify ProblemConstraint where
   simplify' (PConstr pid unblock c) = PConstr pid unblock <$> simplify' c
 
+instance Simplify LocalEquation where
+  simplify' (LocalEquation g t u a) = do
+    g' <- simplify' g
+    (t', u', a') <- addContext g' $ simplify (t, u, a)
+    return $ LocalEquation g' t' u' a'
+
 instance Simplify Constraint where
   simplify' (ValueCmp cmp t u v) = do
     (t,u,v) <- simplify' (t,u,v)
@@ -1201,6 +1247,7 @@ instance Simplify Constraint where
   simplify' c@CheckMetaInst{}     = return c
   simplify' (CheckType t)         = CheckType <$> simplify' t
   simplify' (UsableAtModality cc ms mod t) = flip (UsableAtModality cc) mod <$> simplify' ms <*> simplify' t
+  simplify' (RewConstraint e)     = RewConstraint <$> simplify' e
 
 instance Simplify CompareAs where
   simplify' (AsTermsOf a) = AsTermsOf <$> simplify' a
@@ -1361,6 +1408,12 @@ instance (Subst a, Normalise a) => Normalise (Tele a) where
 instance Normalise ProblemConstraint where
   normalise' (PConstr pid unblock c) = PConstr pid unblock <$> normalise' c
 
+instance Normalise LocalEquation where
+  normalise' (LocalEquation g t u a) = do
+    g' <- normalise g
+    (t', u', a') <- addContext g' $ normalise (t, u, a)
+    return $ LocalEquation g' t' u' a'
+
 instance Normalise Constraint where
   normalise' (ValueCmp cmp t u v) = do
     (t,u,v) <- normalise' (t,u,v)
@@ -1387,6 +1440,7 @@ instance Normalise Constraint where
   normalise' c@CheckMetaInst{}     = return c
   normalise' (CheckType t)         = CheckType <$> normalise' t
   normalise' (UsableAtModality cc ms mod t) = flip (UsableAtModality cc) mod <$> normalise' ms <*> normalise' t
+  normalise' (RewConstraint e)     = RewConstraint <$> normalise' e
 
 instance Normalise CompareAs where
   normalise' (AsTermsOf a) = AsTermsOf <$> normalise' a
@@ -1573,7 +1627,7 @@ instance (Subst a, InstantiateFull a) => InstantiateFull (Abs a) where
     instantiateFull' a@(Abs x _) = Abs x <$> underAbstraction_ a instantiateFull'
     instantiateFull' (NoAbs x a) = NoAbs x <$> instantiateFull' a
 
-instance (InstantiateFull t, InstantiateFull e) => InstantiateFull (Dom' t e) where
+instance InstantiateFull e => InstantiateFull (Dom e) where
     instantiateFull' (Dom i n b tac rew x) = Dom i n b <$> instantiateFull' tac <*> instantiateFull' rew <*> instantiateFull' x
 
 instance InstantiateFull Context where
@@ -1635,6 +1689,7 @@ instance InstantiateFull Constraint where
     c@CheckMetaInst{}   -> return c
     CheckType t         -> CheckType <$> instantiateFull' t
     UsableAtModality cc ms mod t -> flip (UsableAtModality cc) mod <$> instantiateFull' ms <*> instantiateFull' t
+    RewConstraint e     -> RewConstraint <$> instantiateFull' e
 
 instance InstantiateFull CompareAs where
   instantiateFull' (AsTermsOf a) = AsTermsOf <$> instantiateFull' a
@@ -1693,14 +1748,28 @@ instance InstantiateFull GlobalRewriteRule where
       <*> pure c
       <*> pure top
 
-instance InstantiateFull RewriteRule where
-  instantiateFull' (RewriteRule a b c d e) =
-    RewriteRule
+instance InstantiateFull LocalEquation where
+  instantiateFull' (LocalEquation a b c d) =
+    LocalEquation
       <$> instantiateFull' a
       <*> instantiateFull' b
       <*> instantiateFull' c
       <*> instantiateFull' d
+
+instance InstantiateFull RewriteRule where
+  instantiateFull' (RewriteRule a b c d e) =
+    RewriteRule
+      <$> instantiateFull' a
+      <*> pure b
+      <*> instantiateFull' c
+      <*> instantiateFull' d
       <*> instantiateFull' e
+
+instance InstantiateFull RewDom where
+  instantiateFull' (RewDom a b) =
+    RewDom
+      <$> instantiateFull' a
+      <*> instantiateFull' b
 
 instance InstantiateFull DisplayForm where
   instantiateFull' (Display n ps v) = uncurry (Display n) <$> instantiateFull' (ps, v)